+++ rs/index/src/segment/immutable_segment.rs
@@ -46,7 +46,7 @@ impl<Q: Quantizer> Searchable for ImmutableSegment<Q> {
 
     fn search_with_id(
         &self,
-        id: u64,
+        id: u128,
         query: &[f32],
         k: usize,
         ef_construction: u32,
